Fixed bug logout & email verification
